+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
 
+## [1.0.2] - 2016-02-29
+### Changed
+- Rolled own PDepend generator instead of converting
+
+### Fixed
+- Fixed 'he', which was converted incorrectly
+- Fixed project-wide 'i', whose value kept adding up
+
+
 ## [1.0.1] - 2016-02-29
 ### Fixed
 - Conditionally show success/error msg
@@ -16,3 +25,4 @@
 
 [1.0.0]: https://github.com/cauditor/php-analyzer/compare/cdcffeec68ccee59efdee5dd056ea5456b6e4b09...1.0.0
 [1.0.1]: https://github.com/cauditor/php-analyzer/compare/1.0.0...1.0.1
+[1.0.2]: https://github.com/cauditor/php-analyzer/compare/1.0.1...1.0.2